refactor privacy fragment
