+++ app/src/main/res/layout/fragment_privacy.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="24dp"
-    tools:context=".PrivacyFragment">
+    tools:context=".views.privacy.PrivacyFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
